journal: Use size_t instead of unsigned for array sizes

+++ src/libsystemd/sd-journal/journal-file.h
@@ -216,7 +216,7 @@ int journal_file_append_entry(
                 JournalFile *f,
                 const dual_timestamp *ts,
                 const sd_id128_t *boot_id,
-                const struct iovec iovec[], unsigned n_iovec,
+                const struct iovec iovec[], size_t n_iovec,
                 uint64_t *seqno,
                 Object **ret,
                 uint64_t *offset);
